+++ src/commands/index.ts
@@ -1,8 +1,7 @@
 /**
  * Processing-vscode - Processing Language Support for VSCode
  *
- * @version 2.0.3
- * @copyright (C) 2016 - 2020 Tobiah Zarlez, 2021 Luke Zhang
+ * @copyright (C) 2021 Luke Zhang
  */
 
 import {openDocumentation, openProcessingDocs, searchUnityDocs} from "./search"
@@ -11,6 +10,12 @@ import vscode from "vscode"
 
 export const subscribeCommands = (context: vscode.ExtensionContext): void => {
     context.subscriptions.push(vscode.commands.registerCommand("processing.Run", runProject))
+    context.subscriptions.push(
+        vscode.commands.registerCommand("processing.RunPy", () => runProject("py")),
+    )
+    context.subscriptions.push(
+        vscode.commands.registerCommand("processing.RunJava", () => runProject("java")),
+    )
 
     context.subscriptions.push(
         vscode.commands.registerCommand(