+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.0] - 2021-05-21
+
+### Added
+
+-   feat: add python run button
+
+### Changed
+
+-   fix(docs): trim documentation hover data
+-   fix: don't show hover docs on single line comments
+
 ## [2.1.0] - 2021-05-02
 
 ### Added